requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==20.1.0
